pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "basic-scraper"
version = "0.1.0"
description = "Basic scraper: PDFs, text, and images from websites at high quality, stored locally."
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.10"
dependencies = [
    "httpx",
    "beautifulsoup4",
    "lxml",
    "readability-lxml",
]

[project.optional-dependencies]
readability = ["readability-lxml"]
js = ["playwright"]
bundle = ["pyinstaller>=6.0"]

[project.scripts]
scrape = "web_scraper.cli:main"
scrape-gui = "web_scraper.gui:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["web_scraper*"]
